backend/requirements.txt
# Python Backend Dependencies for SmartEstate AI Features
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
openai==1.54.0
pydantic==2.10.0
httpx==0.27.0
googlemaps==4.10.0
geopy==2.4.1
postgrest==0.16.11
python-multipart==0.0.6
redis==5.0.1
